+++ src/main/ipc/init-ipc.ts
@@ -1,6 +1,7 @@
 import { initAthleteHandlers } from "./athletes-ipc";
 import { initdbSettingsHandlers } from "./dbsettings-ipc";
 import { initEventLogsHandlers } from "./eventLogs-ipc";
+import { initExportHandlers } from "./export-ipc";
 import { initFooterHandlers } from "./footer-ipc";
 import { initResourceHandlers } from "./resource-ipc";
 import { initRunnerFormHandlers } from "./runnerform-ipc";
@@ -10,6 +11,7 @@ import { initStatsHandlers } from "./stats-ipc";
 export function initializeIpcHandlers() {
   initAthleteHandlers();
   initdbSettingsHandlers();
+  initExportHandlers();
   initFooterHandlers();
   initResourceHandlers();
   initEventLogsHandlers();
